further refacotr before context addition
